new commit

+++ src/BmiService.java
@@ -1,5 +1,5 @@
 public class BmiService {
-    public double count(double weight, double height) {
+    public int count(double weight, double height) {
 
         double imt = height / (weight * weight);
         return (int) imt;
